Real Fix: Fixed infinite recursion when using image placeholder
If by any reason an image is not found on the server it should turn to the placeholder (default magento one). But how that placeholder is used when the function imagePath() is called again the code generating the new base path is wrong as it only works as long as the image exists in the media folder (while the default placeholder is stored in the skin parent folder) this resulting in an endless loop. I also thought i would be better to first try and use the placeholder set up by the store admin.